cherry-pick:关闭像素取整

+++ frameworks/core/components_ng/base/view_abstract_model_ng.h
@@ -506,7 +506,7 @@ public:
         ViewAbstract::SetLayoutWeight(value);
     }
 
-    void SetPixelRound(uint8_t value) override
+    void SetPixelRound(uint16_t value) override
     {
         ViewAbstract::SetPixelRound(value);
     }
