+++ frameworks/core/components_ng/render/render_context.h
@@ -333,7 +333,7 @@ public:
         return {};
     }
 
-    virtual void SavePaintRect(bool isRound = true, uint8_t flag = 0) {}
+    virtual void SavePaintRect(bool isRound = true, uint16_t flag = 0) {}
     virtual void SyncPartialRsProperties() {}
     virtual void UpdatePaintRect(const RectF& paintRect) {}
